requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.12.0
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==4.5.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
lxml==6.0.0
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.33.0
orjson==3.11.1
packaging==23.2
primp==0.15.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.42
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
